csl_dict renamed to csl_item
